Eagle Has Landed
rewired entire mechanical system of mod. Game starts with single moon. Players can build a Refracting Telescope to "look at the moon" via a gizmo button on the building's ui. This button will produce a map on the moon. Players can build a moon rocket and moon launcher that gives them the fuel needed to make it to the moon. Telescope has Dev option to create new moon which can be viewed to create another map.
NOTE: Known error with biome when making moon map, tile is set to last time made in game which is usually an ocean tile. this creates biome error